Maj precision + verif algorithmes
La précision n'est plus prise en compte sur cette version, car les doucement sont petits. La dissimilarité va de 0 à 1 (au contraire d'avant).

+++ src/Dissimilarity/Dissimilarity.java
@@ -7,5 +7,5 @@ import Execution.Word;
 
 
 public interface Dissimilarity {
-	public int getValue(List<Word> list1, List<Word> list2, int precision);
+	public double getValue(List<Word> list1, List<Word> list2, int precision);
 }
